+++ api/src/main/java/org/testifyproject/annotation/VirtualResource.java
@@ -23,6 +23,7 @@ import java.lang.annotation.Retention;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import java.lang.annotation.Target;
 import java.util.concurrent.TimeUnit;
+import org.testifyproject.DataProvider;
 import org.testifyproject.VirtualResourceProvider;
 
 /**
@@ -94,6 +95,24 @@ public @interface VirtualResource {
      * @return the configKey section key.
      */
     String configKey() default "";
+
+    /**
+     * A list of classpath data files that should be loaded by the virtual
+     * resource prior to being used. Note that
+     * {@link java.nio.file.FileSystem#getPathMatcher(java.lang.String)} glob
+     * patterns are supported
+     *
+     * @return an array of data file names or glob patterns.
+     */
+    String[] dataFiles() default {};
+
+    /**
+     * Specifies a data provider implementations that loads data into the
+     * resource prior to it being used.
+     *
+     * @return the data provider implementation class.
+     */
+    Class<? extends DataProvider> dataProvider() default DataProvider.class;
 
     /**
      * <p>