+++ core/src/test/java/org/testifyproject/core/DefaultRemoteResourceProviderTest.java
@@ -15,6 +15,7 @@
  */
 package org.testifyproject.core;
 
+import java.nio.file.Path;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -29,6 +30,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
+import org.testifyproject.DataProvider;
 import org.testifyproject.Instance;
 import org.testifyproject.RemoteResourceInstance;
 import org.testifyproject.RemoteResourceProvider;
@@ -38,7 +40,9 @@ import org.testifyproject.TestConfigurer;
 import org.testifyproject.TestContext;
 import org.testifyproject.TestDescriptor;
 import org.testifyproject.annotation.RemoteResource;
+import org.testifyproject.core.util.FileSystemUtil;
 import org.testifyproject.core.util.ReflectionUtil;
+import org.testifyproject.fixture.resource.TestDataProvider;
 import org.testifyproject.guava.common.collect.ImmutableList;
 import org.testifyproject.trait.PropertiesReader;
 
@@ -50,14 +54,16 @@ public class DefaultRemoteResourceProviderTest {
 
     DefaultRemoteResourceProvider sut;
     ReflectionUtil reflectionUtil;
+    FileSystemUtil fileSystemUtil;
     List<ResourceInstance<RemoteResource, RemoteResourceProvider, RemoteResourceInstance>> resourceInstances;
 
     @Before
     public void init() {
         reflectionUtil = mock(ReflectionUtil.class);
+        fileSystemUtil = mock(FileSystemUtil.class);
         resourceInstances = mock(List.class, delegatesTo(new LinkedList<>()));
 
-        sut = spy(new DefaultRemoteResourceProvider(reflectionUtil, resourceInstances));
+        sut = spy(new DefaultRemoteResourceProvider(reflectionUtil, fileSystemUtil, resourceInstances));
     }
 
     @Test
@@ -106,6 +112,10 @@ public class DefaultRemoteResourceProviderTest {
         Object configuration = mock(Object.class);
         TestConfigurer testConfigurer = mock(TestConfigurer.class);
         RemoteResourceInstance<Object> remoteResourceInstance = mock(RemoteResourceInstance.class);
+        String[] dataFilePatterns = {"test.class"};
+        List<Path> dataFiles = mock(List.class);
+        Class dataProviderType = TestDataProvider.class;
+        DataProvider dataProvider = mock(TestDataProvider.class);
         String fqn = "fqn";
         Map<String, Object> properties = mock(Map.class);
 
@@ -126,6 +136,10 @@ public class DefaultRemoteResourceProviderTest {
         given(remoteResourceProvider.configure(testContext, remoteResource, configReader)).willReturn(configuration);
         given(testConfigurer.configure(testContext, configuration)).willReturn(configuration);
         given(remoteResourceProvider.start(testContext, remoteResource, configuration)).willReturn(remoteResourceInstance);
+        given(remoteResource.dataFiles()).willReturn(dataFilePatterns);
+        given(fileSystemUtil.findClasspathFiles(dataFilePatterns)).willReturn(dataFiles);
+        given(remoteResource.dataProvider()).willReturn(dataProviderType);
+        given(reflectionUtil.newInstance(dataProviderType)).willReturn(dataProvider);
         given(remoteResourceInstance.getFqn()).willReturn(fqn);
         given(remoteResourceInstance.getProperties()).willReturn(properties);
         willDoNothing().given(sut).processInstance(remoteResource, remoteResourceInstance, value, serviceInstance);
@@ -143,6 +157,13 @@ public class DefaultRemoteResourceProviderTest {
         verify(remoteResourceProvider).configure(testContext, remoteResource, configReader);
         verify(testConfigurer).configure(testContext, configuration);
         verify(remoteResourceProvider).start(testContext, remoteResource, configuration);
+        verify(remoteResource).dataFiles();
+        verify(fileSystemUtil).findClasspathFiles(dataFilePatterns);
+        verify(remoteResourceProvider).load(testContext, remoteResource, remoteResourceInstance, dataFiles);
+        verify(remoteResource).dataProvider();
+        verify(reflectionUtil).newInstance(dataProviderType);
+        verify(serviceInstance).inject(dataProvider);
+        verify(dataProvider).load(testContext, dataFiles, remoteResourceInstance);
         verify(remoteResourceInstance).getFqn();
         verify(remoteResourceInstance).getProperties();
         verify(testContext).addProperty(fqn, properties);